requirements.txt
line-bot-sdk==3.9.0
flask==3.0.0
gspread==6.0.0
google-auth==2.25.2
google-api-python-client==2.110.0
google-cloud-speech==2.23.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
gunicorn==21.2.0
pydub==0.25.1
pandas==2.1.4
matplotlib==3.8.2
redis==5.0.1
APScheduler==3.10.4
jieba==0.42.1
pytest==7.4.3
pytest-mock==3.12.0
cryptography==41.0.8